pyproject.toml
[project]
name = "awair"
version = "0.0.2"
description = "Awair API client and data collection system with AWS Lambda automation"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Ryan Williams", email = "ryan@runsascoded.com"}
]
maintainers = [
    {name = "Ryan Williams", email = "ryan@runsascoded.com"}
]
keywords = ["awair", "air-quality", "iot", "sensors", "aws", "lambda", "data-collection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "pyarrow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/runsascoded/awair"
Documentation = "https://github.com/runsascoded/awair#readme"
Repository = "https://github.com/runsascoded/awair.git"
Issues = "https://github.com/runsascoded/awair/issues"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]
lambda = [
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
    "utz>=0.20.0",
]
all = [
    "ruff",
    "pytest",
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
    "utz>=0.20.0",
]
[project.scripts]
awair = "awair.cli.base:awair"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 999
target-version = "py39"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
]
ignore = [
    "E501",  # line too long (handled by formatter)
]
